Functions

Problem

Sparspak.SpkProblem.infullrhs — Method

Sparspak.SpkProblem.makegridproblem — Method

This routine constructs a Grid object given an H and K, and fills in a Problem object using this Grid. Input Parameters: h - the number of rows in the Grid k - the number of columns in the Grid stencil - an optional variable specifying the difference operator to be applied to the grid. Output Parameter: p - the Problem object to be filled

Sparspak.SpkProblem.makegridproblem — Method

This routine fills in a Problem object using a given Grid. Input Parameters: g - the Grid to be used to fill a Problem matrix stencil - an optional variable specifying the difference operator to be applied to the grid. Output Parameter: p - the Problem object to be filled

Sparspak.SpkProblem.makerhs — Method

This routine constructs the RHS of a problem given an x for the equation $Ax = rhs"". The x must have the same number of elements as the problem (represented by A above) has columns. If x is not present, a right hand side is contructed so that (a, the) solution is 1, 2, 3, ...m. Input Parameter: x - the vector in the equation$Ax = rhs"" mType - matrix type (optional). If the matrix is symmetric and only the lower or upper triangle is present, the user must let the routine know this by setting mType to one of: "L" or "l" - when only the lower triangle is present "U" or "u" - when only the upper triangle is present "T" or "t" - when either the lower or upper triangle is present. Updated Parameter: p - the problem for which the RHS is being constructed.

Sparse LU Method

Sparspak.SpkSparseBase.factor — Method

factor(s::SparseSolver{IT}) where {IT}

Numerical factorization of the coefficient matrix.

Sparspak.SpkSparseBase.findorder — Method

findorder(s::SparseSolver{IT}, orderfunction::F) where {IT, F}

Find reordering of the coefficient matrix.

Sparspak.SpkSparseBase.findorder — Method

findorder(s::SparseSolver{IT}) where {IT, F}

Find reordering of the coefficient matrix using the default method.

Sparspak.SpkSparseBase.inmatrix — Method

inmatrix(s::SparseSolver{IT}, p::Problem{IT}) where {IT}

Put numerical values of the matrix stored in the problem into the data structures of the solver.

Sparspak.SpkSparseBase.symbolicfactor — Method

symbolicfactor(s::SparseSolver{IT})

Symbolic factorization of the(reordered) matrix A and the creation of the data structures for the factorization and forward and backward substitution.

Sparspak.SpkSparseBase.triangularsolve — Method

triangularsolve(s::SparseSolver{IT, FT}, solution::Vector{FT}) where {IT, FT}

Forward and backward substitution (triangular solution).

Variant where the right-hand side vector is passed in.

Sparspak.SpkSparseBase.triangularsolve — Method

triangularsolve(s::SparseSolver{IT},  p::Problem{IT}) where {IT}

Forward and backward substitution (triangular solution).

Sparspak.SpkSparseSolver.findorderperm — Method

findorderperm(s::SparseSolver{IT}, perm) where {IT}

Find reordering of the coefficient matrix using a given permutation.

Sparspak.SpkSparseSolver.solve — Method

solve(s::SparseSolver{IT}, p::Problem{IT}) where {IT}

Execute all the steps of the solution process:

Reordering of the matrix A

Symbolic factorization of the(reordered) matrix A and the creation of the

data structures for the factorization and forward and backward substitution

Putting numerical values of

A into the data structures

Numerical factorization of A

Forward and backward substitution (triangular solution)

Graphs

Sparspak.SpkGraph.sortgraph — Method

SortGraph - sort the adjacency lists of the graph Important assumption: This works only for graphs that are symmetric. Output: updated graph

Ordering

Multiple minimum degree (MMD) ordering.

Sparspak.SpkMmd.generalmmd — Method

generalmmd(n, xadj, adj, perm, invp)

This routine implements the minimum degree algorithm. It makes use of the implicit representation of elimination graphs by quotient graphs, and the notion of indistinguishable nodes. It also implements the modifications by multiple elimination and minimum external degree.

Input parameters -

n - number of equations (xadj, adj) - adjacency structure for the graph. delta - tolerance value for multiple elimination. FIX ME: should delta be passed as argument?

Output: perm, invp - the minimum degree Ordering.

Working arrays - degHead (deg) - points to first node with degree deg, or 0 if there are no such nodes. degNext (node) - points to the next node in the degree list associated with node, or 0 if node was the last in the degree list. degPrev (node) - points to the previous node in a degree list associated with node, or the negative of the degree of node (if node was the last in the degree list), or 0 if the node is not in the degree lists. superSIze - the size of the supernodes. elimHead - points to the first node eliminated in the current pass Using elimNext, one can determine all nodes just eliminated. elimNext (node) - points to the next node in a eliminated supernode or 0 if there are no more after node. marker - marker vector. mergeParent - the parent map for the merged forest. needsUpdate (node) - > 0 iff node needs degree update.(0 otherwise)

Sparspak.SpkMmd.mmdelim — Method

Purpose - This routine eliminates the node mdNode of
  minimum degree from the adjacency structure, which
  is stored in the quotient Graph format.  It also
  transforms the quotient Graph representation of the
  elimination Graph.

Input parameters - mdNode - node of minimum degree. tag - tag value. invp - the inverse of an incomplete minimum degree Ordering. (It is zero at positions where the Ordering is unknown.) Updated parameters - (xadj, adjncy) - updated adjacency structure (xadj is not updated). degHead (deg) - points to first node with degree deg, or 0 if there are no such nodes. degNext (node) - points to the next node in the degree list associated with node, or 0 if node was the last in the degree list. degPrev (node) - points to the previous node in a degree list associated with node, or the negative of the degree of node (if node was the last in the degree list), or 0 if the node is not in the degree lists. superSIze - the size of the supernodes. elimNext (node) - points to the next node in a eliminated supernode or 0 if there are no more after node. marker - marker vector. mergeParent - the parent map for the merged forest. needsUpdate (node) - > 0 iff node needs update. (0 otherwise)

Sparspak.SpkMmd.mmdupdate — Method

Purpose - This routine updates the degrees of nodes
  after a multiple elimination step.

Input parameters - elimHead - the beginning of the list of eliminated nodes (i.e., newly formed elements). neqns - number of equations. (xadj, adjncy) - adjacency structure. delta - tolerance value for multiple elimination. invp - the inverse of an incomplete minimum degree Ordering. (It is zero at positions where the Ordering is unknown.) Updated parameters - mindeg - new minimum degree after degree update. degHead (deg) - points to first node with degree deg, or 0 if there are no such nodes. degNext (node) - points to the next node in the degree list associated with node, or 0 if node was the last in the degree list. degPrev (node) - points to the previous node in a degree list associated with node, or the negative of the degree of node (if node was the last in the degree list), or 0 if the node is not in the degree lists. superSIze - the size of the supernodes. elimNext (node) - points to the next node in a eliminated supernode or 0 if there are no more after node. marker - marker vector for degree update. tag - tag value. mergeParent - the parent map for the merged forest. needsUpdate (node) - > 0 iff node needs update. (0 otherwise)

Elimination Trees

Sparspak.SpkETree.binarytree — Method

Sparspak.SpkETree.findetree — Method

Sparspak.SpkETree.getetree — Method

Given a graph and an ordering, GetETree finds the corresponding elimination tree. It calls the subroutine FindETree, which actually does the work.

Input Parameter: g - the graph whose elimination tree is to be found. order - the ordering for g

Updated Parameters: t - the elimination tree.

Sparspak.SpkETree.getpostorder — Method

Sparspak.SpkETree.postordertree — Method

Sparspak.SpkETree.weightedbinarytree — Method

Symbolic Factorization

Sparspak.SpkSymfct.findcolumncounts — Method

This subroutine determines the column counts in
the Cholesky factor.  It uses an algorithm due to Joseph Liu
found in SIMAX 11, 1990, pages 144 - 145.0

Input parameters: (i) n - number of equations. (i) xadj - array of length n + 1, containing pointers to the adjacency structure. (i) adj - array of length xadj(n + 1) - 1, containing the adjacency structure. (i) perm - array of length n, containing the postordering. (i) invp - array of length n, containing the inverse of the postordering. (i) parent - array of length n, containing the elimination tree of the postordered matrix.

Output parameters: (i) colcnt - array of length n, containing the number of nonzeros in each column of the factor, including the diagonal entry. (i) nlnz - number of nonzeros in the factor, including the diagonal entries.

Work parameters: (i) marker - array of length n used to mark the vertices visited in each row subtree.

Sparspak.SpkSymfct.findnumberofsupermods — Method

Sparspak.SpkSymfct.findschedule — Method

Sparspak.SpkSymfct.findsupernodes — Method

Sparspak.SpkSymfct.findsupernodetree — Method

Sparspak.SpkSymfct.symbolicfact — Method

This routine performs supernodal symbolic
factorization on a reordered linear system.
This is essentially a Fortran 90 translation of a code written
by Esmond Ng and Barry Peyton.

Input parameters: (i) n - number of equations (i) xadj - array of length n + 1 containing pointers to the adjacency structure. (i) adj - array of length xadj(n + 1) - 1 containing the adjacency structure. (i) perm - array of length n containing the postordering. (i) invp - array of length n containing the inverse of the postordering. (i) colcnt - array of length n, containing the number of nonzeros (non - empty rows) in each column of the factor, including the diagonal entry. (i) nsuper - number of supernodes. (i) xsuper - array of length nsuper + 1, containing the first column of each supernode. (i) snode - array of length n for recording supernode membership. (i) nofsub - number of subscripts to be stored in lindx.

Output parameters: (i) xlindx - array of length n + 1, containing pointers into the subscript vector. (i) lindx - array of length maxsub, containing the compressed subscripts. (i) xlnz - column pointers for l.

Working parameters: (i) mrglnk - array of length nsuper, containing the children of each supernode as a linked list. (i) rchlnk - array of length n + 1, containing the current linked list of merged indices (the "reach" set). (i) marker - array of length n used to mark indices as they are introduced into each supernode"s index set.

Grid