Types

Problem

Sparspak.SpkProblem.Problem — Type

Sparse LU Solver

Missing docstring.

Missing docstring for SparseSolver. Check Documenter's build log for details.

Elimination Trees

Sparspak.SpkETree.ETree — Type

Graphs

Sparspak.SpkGraph.Graph — Type

This routine constructs a graph from a problem object.

It does not check that the problem object contains a structurally symmetric matrix, since sometimes only the lower or upper triangle of a symmetric matrix may be stored. There are routines in this module to make a given graph object structurally symmetric.

Input: g - the graph object, declared by the calling routine p - the problem object, used to create the graph diagonal - indicates that the diagonal elements are included. If diagonal is not given, the adjacency structure does not include the diagonal elements. objectName - (optional) name to be assigned to g. Updated Parameter: g - created graph object.

Ordering

Sparspak.SpkOrdering.Ordering — Type

ConstructOrdering constructs an ordering object. Since only one parameter (nRows) is supplied, it is assumed that the size of the row ordering and column ordering are the same. That is, that the matrix is square. Input Parameters: order - the ordering (declared in the calling program) nRows - the number of rows (and columns) in the matrix objectName - the name of the ordering object (optional) Output Parameter: order - the updated ordering object

ConstructOrdering2 constructs an ordering object. The arrays rPerm, cPerm, rInvp, cInvp are allocated and initialized to the identity permutation. Input Parameter: order - the ordering (declared in the calling program) nRows, nCols - the number of rows and columns in the matrix objectName - the name of the ordering object (optional) Output Parameter: order - the updated ordering object

Grid

Sparspak.SpkGrid.Grid — Type

Functions

Problem

Missing docstring.

Missing docstring for Problem(nrows::IT, ncols::IT, nnz::IT=2500, z::FT=0.0, info = "") where {IT, FT}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for inaij!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for inbi!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for insparse!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for infullrhs!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for outsparse. Check Documenter's build log for details.

Missing docstring.

Missing docstring for computeresidual. Check Documenter's build log for details.

Sparspak.SpkProblem.makerhs — Function

This routine constructs the RHS of a problem given an x for the equation $Ax = rhs"". The x must have the same number of elements as the problem (represented by A above) has columns. If x is not present, a right hand side is contructed so that (a, the) solution is 1, 2, 3, ...m. Input Parameter: x - the vector in the equation$Ax = rhs"" mType - matrix type (optional). If the matrix is symmetric and only the lower or upper triangle is present, the user must let the routine know this by setting mType to one of: "L" or "l" - when only the lower triangle is present "U" or "u" - when only the upper triangle is present "T" or "t" - when either the lower or upper triangle is present. Updated Parameter: p - the problem for which the RHS is being constructed.

Missing docstring.

Missing docstring for makegridproblem. Check Documenter's build log for details.

Sparse LU Solver

Missing docstring.

Missing docstring for SparseSolver. Check Documenter's build log for details.

Missing docstring.

Missing docstring for findorder!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for findorderperm!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for symbolicfactor!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for inmatrix!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for factor!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for triangularsolve!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for solve!. Check Documenter's build log for details.

Graphs

Modules = [Sparspak.SpkGraph]
Private = true
Order = [:function]

Ordering

Modules = [Sparspak.SpkOrdering]
Private = true
Order = [:function]

Multiple minimum degree (MMD) ordering.

Modules = [Sparspak.SpkMmd]
Private = true
Order = [:function]

Elimination Trees

Modules = [Sparspak.SpkETree]
Private = true
Order = [:function]

Symbolic Factorization

Modules = [Sparspak.SpkSymfct]
Private = true
Order = [:function]

Grid

Modules = [Sparspak.SpkGrid]
Private = true
Order = [:function]