Types

Problem

Sparspak.SpkProblem.Problem — Method

Sparse LU Method

Sparse LU SPD Method

Functions for Symmetric Positive Definite (SPD) matrices.

Elimination Trees

Sparspak.SpkETree.ETree — Method

Graphs

Sparspak.SpkGraph.Graph — Method

This routine constructs a graph from a problem object.

It does not check that the problem object contains a structurally symmetric matrix, since sometimes only the lower or upper triangle of a symmetric matrix may be stored. There are routines in this module to make a given graph object structurally symmetric.

Input: g - the graph object, declared by the calling routine p - the problem object, used to create the graph diagonal - indicates that the diagonal elements are included. If diagonal is not given, the adjacency structure does not include the diagonal elements. objectName - (optional) name to be assigned to g. Updated Parameter: g - created graph object.

Ordering

Sparspak.SpkOrdering.Ordering — Method

ConstructOrdering constructs an ordering object. Since only one parameter (nRows) is supplied, it is assumed that the size of the row ordering and column ordering are the same. That is, that the matrix is square. Input Parameters: order - the ordering (declared in the calling program) nRows - the number of rows (and columns) in the matrix objectName - the name of the ordering object (optional) Output Parameter: order - the updated ordering object

Sparspak.SpkOrdering.Ordering — Method

ConstructOrdering2 constructs an ordering object. The arrays rPerm, cPerm, rInvp, cInvp are allocated and initialized to the identity permutation. Input Parameter: order - the ordering (declared in the calling program) nRows, nCols - the number of rows and columns in the matrix objectName - the name of the ordering object (optional) Output Parameter: order - the updated ordering object

Grid

Sparspak.SpkGrid.Grid — Method