Functions

Problem

Sparspak.SpkProblem.Problem — Method

Missing docstring.

Missing docstring for inaij!(p::Problem{IT,FT}, rnum, cnum, aij=zero(FT)) where {IT,FT}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for inbi!(p::Problem{IT, FT}, rnum::IT, bi::FT) where {IT, FT}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for insparse!(p::Problem{IT,FT}, spm) where {IT,FT}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for infullrhs!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for outsparse. Check Documenter's build log for details.

Missing docstring.

Missing docstring for computeresidual. Check Documenter's build log for details.

Missing docstring.

Missing docstring for makerhs!. Check Documenter's build log for details.

Sparspak.SpkProblem.makegridproblem — Function

This routine fills in a Problem object using a given Grid. Input Parameters: g - the Grid to be used to fill a Problem matrix stencil - an optional variable specifying the difference operator to be applied to the grid. Output Parameter: p - the Problem object to be filled

This routine constructs a Grid object given an H and K, and fills in a Problem object using this Grid. Input Parameters: h - the number of rows in the Grid k - the number of columns in the Grid stencil - an optional variable specifying the difference operator to be applied to the grid. Output Parameter: p - the Problem object to be filled

Sparse LU Solver

Missing docstring.

Missing docstring for SparseSolver. Check Documenter's build log for details.

Missing docstring.

Missing docstring for findorder!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for findorderperm!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for symbolicfactor!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for inmatrix!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for factor!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for triangularsolve!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for solve!. Check Documenter's build log for details.

Graphs

Sparspak.SpkGraph.Graph — Method

This routine constructs a graph from a problem object.

It does not check that the problem object contains a structurally symmetric matrix, since sometimes only the lower or upper triangle of a symmetric matrix may be stored. There are routines in this module to make a given graph object structurally symmetric.

Input: g - the graph object, declared by the calling routine p - the problem object, used to create the graph diagonal - indicates that the diagonal elements are included. If diagonal is not given, the adjacency structure does not include the diagonal elements. objectName - (optional) name to be assigned to g. Updated Parameter: g - created graph object.

Missing docstring.

Missing docstring for makestructuresymmetric(g::Graph{IT}) where {IT}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for sortgraph!(g::Graph{IT}) where {IT}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for isstructuresymmetric(g::Graph{IT}) where {IT}. Check Documenter's build log for details.

Ordering

Multiple minimum degree (MMD) ordering.

Elimination Trees

Symbolic Factorization

Grid