Functions

Problem

Sparspak.SpkProblem.Problem — Method

Problem(nrows::IT, ncols::IT, nnz::IT=2500, z::FT=0.0, info = "") where {IT, FT}

Construct a problem.

Sparspak.SpkProblem.inaij! — Function

inaij!(p::Problem{IT,FT}, rnum, cnum, aij=zero(FT)) where {IT,FT}

Input a matrix coefficient.

The value is added to the existing contents.

Sparspak.SpkProblem.inbi! — Function

inbi!(p::Problem{IT, FT}, rnum::IT, bi::FT) where {IT, FT}

Input an entry of the right-hand side vector.

Sparspak.SpkProblem.insparse! — Function

insparse(p::Problem{IT,FT}, spm) where {IT,FT}

Input sparse matrix.

Build a problem from a sparse matrix.

insparse(p::Problem{IT,FT}, I::Vector{IT}, J::Vector{IT}, V::Vector{FT}) where 
{IT,FT}

Build a problem from a sparse matrix in the COO format.

Sparspak.SpkProblem.infullrhs! — Function

infullrhs!(p::Problem{IT,FT}, rhs)  where {IT,FT}

InRHSProblem adds a vector of values, rhs, to the current right hand side of a problem object.

Input:

rhs - the source right-hand side. It must be of length at least p.nrows and if it is greater than p.nrows, only the first p.nrows are used.

Updated:

p - the problem in which rhs is to be inserted.

Sparspak.SpkProblem.outsparse — Function

outsparse(p::Problem{IT,FT})  where {IT,FT}

Output the sparse matrix.

Sparspak.SpkProblem.computeresidual — Function

computeresidual(p::Problem, res::Vector{FT}, xin::Vector{FT} = FT[], mtype = "T") where {FT}

Compute the residual of a problem.

Given a vector x, this routine calculates the difference between the RHS of the given Problem and A*x and places this in res.

Input:

p - the Problem used to calculate res, using xin

xin - the input "solution" vector used to compute the residual

mtype - matrix type (optional). If the matrix is symmetric and only the lower or upper triangle is present, the user must let the routine know this by setting mType to one of: "L" or "l" - when only the lower triangle is present "U" or "u" - when only the upper triangle is present "T" or "t" - when either the lower or upper triangle is present.

Output: res - the calculated residual

Missing docstring.

Missing docstring for makerhs. Check Documenter's build log for details.

Sparspak.SpkProblem.makegridproblem — Function

makegridproblem(g::Grid{IT}) where {IT}

This routine fills in a problem object using a given grid.

Input:

g - the Grid to be used to fill a Problem matrix

Output:

p - the Problem object to be filled

makegridproblem(h::IT, k::IT) where {IT}

Construct a problem object based on a grid.

Input:

h - the number of rows in the Grid

k - the number of columns in the Grid

Output:

p - the Problem object to be filled

Sparse LU Solver

Sparspak.SpkSparseSolver.SparseSolver — Type

SparseSolver{IT, FT}

Type of LU general sparse solver.

Sparspak.SpkSparseSolver.findorder! — Function

findorder!(s::SparseSolver{IT}, orderfunction::F) where {IT, F}

Find reordering of the coefficient matrix.

findorder!(s::SparseSolver{IT}) where {IT, F}

Find reordering of the coefficient matrix using the default method.

Sparspak.SpkSparseSolver.findorderperm! — Function

findorderperm!(s::SparseSolver{IT}, perm) where {IT}

Find reordering of the coefficient matrix using a given permutation.

Sparspak.SpkSparseSolver.symbolicfactor! — Function

symbolicfactor!(s::SparseSolver{IT})

Symbolic factorization of the(reordered) matrix A.

Create the data structures for the factorization and forward and backward substitution.

Sparspak.SpkSparseSolver.inmatrix! — Function

inmatrix!(s::SparseSolver{IT}, p::Problem{IT}) where {IT}

Put numerical values of the matrix stored in the problem into the data structures of the solver.

Sparspak.SpkSparseSolver.factor! — Function

factor!(s::SparseSolver{IT}) where {IT}

Numerical factorization of the coefficient matrix.

Sparspak.SpkSparseSolver.triangularsolve! — Function

triangularsolve!(s::SparseSolver{IT},  p::Problem{IT}) where {IT}

Forward and backward substitution (triangular solution).

triangularsolve!(s::SparseSolver{IT, FT}, solution::Vector{FT}) where {IT, FT}

Forward and backward substitution (triangular solution).

Variant where the right-hand side vector is passed in.

Sparspak.SpkSparseSolver.solve! — Function

solve!(s::SparseSolver{IT}, p::Problem{IT}) where {IT}

Execute all the steps of the solution process.

Given a symmetric matrix A, the steps are:

Reordering of the matrix A.

Symbolic factorization of the(reordered) matrix A.

Putting numerical values of A into the data structures.

Numerical factorization of A.

Forward and backward substitution (triangular solution).

The solution can be retrieved as p.x.

Missing docstring.

Missing docstring for Order = [:function]. Check Documenter's build log for details.

Graphs

Missing docstring.

Missing docstring for Modules = [Sparspak.SpkGraph]. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Private = true. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Order = [:function]. Check Documenter's build log for details.

Ordering

Missing docstring.

Missing docstring for Modules = [Sparspak.SpkOrdering]. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Private = true. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Order = [:function]. Check Documenter's build log for details.

Multiple minimum degree (MMD) ordering.

Missing docstring.

Missing docstring for Modules = [Sparspak.SpkMmd]. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Private = true. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Order = [:function]. Check Documenter's build log for details.

Elimination Trees

Missing docstring.

Missing docstring for Modules = [Sparspak.SpkETree]. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Private = true. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Order = [:function]. Check Documenter's build log for details.

Symbolic Factorization

Missing docstring.

Missing docstring for Modules = [Sparspak.SpkSymfct]. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Private = true. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Order = [:function]. Check Documenter's build log for details.

Grid

Missing docstring.

Missing docstring for Modules = [Sparspak.SpkGrid]. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Private = true. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Order = [:function]. Check Documenter's build log for details.