Functions

Problem

Sparspak.SpkProblem.infullrhs — Method

Sparspak.SpkProblem.makegridproblem — Method

This routine constructs a Grid object given an H and K, and fills in a Problem object using this Grid. Input Parameters: h - the number of rows in the Grid k - the number of columns in the Grid stencil - an optional variable specifying the difference operator to be applied to the grid. Output Parameter: p - the Problem object to be filled

Sparspak.SpkProblem.makegridproblem — Method

This routine fills in a Problem object using a given Grid. Input Parameters: g - the Grid to be used to fill a Problem matrix stencil - an optional variable specifying the difference operator to be applied to the grid. Output Parameter: p - the Problem object to be filled

Sparspak.SpkProblem.makerhs — Method

This routine constructs the RHS of a problem given an x for the equation $Ax = rhs"". The x must have the same number of elements as the problem (represented by A above) has columns. If x is not present, a right hand side is contructed so that (a, the) solution is 1, 2, 3, ...m. Input Parameter: x - the vector in the equation$Ax = rhs"" mType - matrix type (optional). If the matrix is symmetric and only the lower or upper triangle is present, the user must let the routine know this by setting mType to one of: "L" or "l" - when only the lower triangle is present "U" or "u" - when only the upper triangle is present "T" or "t" - when either the lower or upper triangle is present. Updated Parameter: p - the problem for which the RHS is being constructed.

Sparspak.SpkSparseBase.factor — Method

factor(s::SparseSolver{IT}) where {IT}

Numerical factorization of the coefficient matrix.

Sparspak.SpkSparseBase.findorder — Method

findorder(s::SparseSolver{IT}, orderfunction::F) where {IT, F}

Find reordering of the coefficient matrix.

Sparspak.SpkSparseBase.findorder — Method

findorder(s::SparseSolver{IT}) where {IT, F}

Find reordering of the coefficient matrix using the default method.

Sparspak.SpkSparseBase.inmatrix — Method

inmatrix(s::SparseSolver{IT}, p::Problem{IT}) where {IT}

Put numerical values of the matrix stored in the problem into the data structures of the solver.

Sparspak.SpkSparseBase.symbolicfactor — Method

symbolicfactor(s::SparseSolver{IT})

Symbolic factorization of the(reordered) matrix A and the creation of the data structures for the factorization and forward and backward substitution.

Sparspak.SpkSparseBase.triangularsolve — Method

triangularsolve(s::SparseSolver{IT, FT}, solution::Vector{FT}) where {IT, FT}

Forward and backward substitution (triangular solution).

Variant where the right-hand side vector is passed in.

Sparspak.SpkSparseBase.triangularsolve — Method

triangularsolve(s::SparseSolver{IT},  p::Problem{IT}) where {IT}

Forward and backward substitution (triangular solution).

Sparspak.SpkSparseSolver.findorderperm — Method

findorderperm(s::SparseSolver{IT}, perm) where {IT}

Find reordering of the coefficient matrix using a given permutation.

Sparspak.SpkSparseSolver.solve — Method

solve(s::SparseSolver{IT}, p::Problem{IT}) where {IT}

Execute all the steps of the solution process:

Reordering of the matrix A

Symbolic factorization of the(reordered) matrix A and the creation of the

data structures for the factorization and forward and backward substitution

Putting numerical values of

A into the data structures

Numerical factorization of A

Forward and backward substitution (triangular solution)