Types

Problem

Sparspak.SpkProblem.Problem — Type

Problem{IT, FT}

Type of a sparse-matrix coupled linear algebraic equations problem.

Sparse LU Solver

Sparspak.SpkSparseSolver.SparseSolver — Type

SparseSolver{IT, FT}

Type of LU general sparse solver.

Elimination Trees

Sparspak.SpkETree.ETree — Type

ETree(nv::IT) where {IT}

Construct elimination tree.

Graphs

Sparspak.SpkGraph.Graph — Type

This routine constructs a graph from a problem object.

It does not check that the problem object contains a structurally symmetric matrix, since sometimes only the lower or upper triangle of a symmetric matrix may be stored. There are routines in this module to make a given graph object structurally symmetric.

Input: g - the graph object, declared by the calling routine p - the problem object, used to create the graph diagonal - indicates that the diagonal elements are included. If diagonal is not given, the adjacency structure does not include the diagonal elements. objectName - (optional) name to be assigned to g. Updated Parameter: g - created graph object.

Ordering

Sparspak.SpkOrdering.Ordering — Type

Ordering{IT}

Type of ordering of the rows and columns.

Grid

Sparspak.SpkGrid.Grid — Type

Grid(h::IT, k::IT) where {IT}

Construct a grid with a given number of spacings.

Functions

Problem

Problem
inaij
inbi
insparse
infullrhs
outsparse
computeresidual
makerhs
makegridproblem

Graphs

Modules = [Sparspak.SpkGraph]
Private = true
Order = [:function]

Ordering

Modules = [Sparspak.SpkOrdering]
Private = true
Order = [:function]

Multiple minimum degree (MMD) ordering.

Modules = [Sparspak.SpkMmd]
Private = true
Order = [:function]

Elimination Trees

Modules = [Sparspak.SpkETree]
Private = true
Order = [:function]

Symbolic Factorization

Modules = [Sparspak.SpkSymfct]
Private = true
Order = [:function]

Grid

Modules = [Sparspak.SpkGrid]
Private = true
Order = [:function]