Types

Problem

Sparspak.SpkProblem.Problem — Type

Problem{IT, FT}

Type of a sparse-matrix coupled linear algebraic equations problem.

Sparspak.SpkProblem.Problem — Method

Problem(nrows::IT, ncols::IT, nnz::IT=2500, z::FT=0.0, info = "") where {IT, FT}

Construct a problem.

Sparse LU Method

Sparspak.SpkSparseSolver.SparseSolver — Type

SparseSolver{IT, FT}

Type of LU general sparse solver.

Sparse LU SPD Method

Functions for Symmetric Positive Definite (SPD) matrices.

Elimination Trees

Sparspak.SpkETree.ETree — Method

ETree(nv::IT) where {IT}

Construct elimination tree.

Graphs

Sparspak.SpkGraph.Graph — Method

This routine constructs a graph from a problem object.

It does not check that the problem object contains a structurally symmetric matrix, since sometimes only the lower or upper triangle of a symmetric matrix may be stored. There are routines in this module to make a given graph object structurally symmetric.

Input: g - the graph object, declared by the calling routine p - the problem object, used to create the graph diagonal - indicates that the diagonal elements are included. If diagonal is not given, the adjacency structure does not include the diagonal elements. objectName - (optional) name to be assigned to g. Updated Parameter: g - created graph object.

Ordering

Sparspak.SpkOrdering.Ordering — Type

Ordering{IT}

Type of ordering of the rows and columns.

Sparspak.SpkOrdering.Ordering — Method

Ordering(nrows::IT) where {IT}

Construct an ordering object.

Since only one parameter(nrows) is supplied, it is assumed that the size of the row ordering and column ordering are the same. That is, that the matrix is square.

Input Parameters: order - the ordering (declared in the calling program) nRows - the number of rows (and columns) in the matrix objectName - the name of the ordering object (optional)

Sparspak.SpkOrdering.Ordering — Method

Ordering(nrows::IT, ncols::IT) where {IT}

Construct an ordering object.

The arrays rperm, cperm, rinvp, cinvp are allocated and initialized to the identity permutation.

Input Parameter: order - the ordering (declared in the calling program) nRows, nCols - the number of rows and columns in the matrix objectName - the name of the ordering object (optional)

Grid

Sparspak.SpkGrid.Grid — Method

Grid(h::IT, k::IT) where {IT}

Construct a grid with a given number of spacings.

Functions

Problem

Sparspak.SpkProblem.computeresidual — Method

computeresidual(p::Problem, res::Vector{FT}, xin::Vector{FT} = FT[], mtype = "T") where {FT}

Compute the residual of a problem.

Given a vector x, this routine calculates the difference between the RHS of the given Problem and A*x and places this in res.

Input:

p - the Problem used to calculate res, using xin

xin - the input "solution" vector used to compute the residual

mtype - matrix type (optional). If the matrix is symmetric and only the lower or upper triangle is present, the user must let the routine know this by setting mType to one of: "L" or "l" - when only the lower triangle is present "U" or "u" - when only the upper triangle is present "T" or "t" - when either the lower or upper triangle is present.

Output: res - the calculated residual

Sparspak.SpkProblem.inaij — Method

inaij(p::Problem{IT,FT}, rnum, cnum, aij=zero(FT)) where {IT,FT}

Input a matrix coefficient.

The value is added to the existing contents.

Sparspak.SpkProblem.inbi — Method

inbi(p::Problem{IT, FT}, rnum::IT, bi::FT) where {IT, FT}

Input an entry of the right-hand side vector.

Sparspak.SpkProblem.infullrhs — Method

infullrhs(p::Problem{IT,FT}, rhs)  where {IT,FT}

InRHSProblem adds a vector of values, rhs, to the current right hand side of a problem object.

Input:

rhs - the source right-hand side. It must be of length at least p.nrows and if it is greater than p.nrows, only the first p.nrows are used.

Updated:

p - the problem in which rhs is to be inserted.

Sparspak.SpkProblem.insparse — Method

insparse(p::Problem{IT,FT}, spm) where {IT,FT}

Input sparse matrix.

Build a problem from a sparse matrix.

Sparspak.SpkProblem.insparse — Method

insparse(p::Problem{IT,FT}, I::Vector{IT}, J::Vector{IT}, V::Vector{FT}) where 
{IT,FT}

Build a problem from a sparse matrix in the COO format.

Sparspak.SpkProblem.makegridproblem — Method

makegridproblem(h::IT, k::IT) where {IT}

Construct a problem object based on a grid.

Input:

h - the number of rows in the Grid

k - the number of columns in the Grid

Output:

p - the Problem object to be filled

Sparspak.SpkProblem.makegridproblem — Method

makegridproblem(g::Grid{IT}) where {IT}

This routine fills in a problem object using a given grid.

Input:

g - the Grid to be used to fill a Problem matrix

Output:

p - the Problem object to be filled

Sparspak.SpkProblem.makerhs — Method

makerhs(p::Problem, x::Vector{FT} = FT[], mtype = "T") where {FT}

This routine constructs the RHS of a problem given an x for the equation Ax = rhs. The x must have the same number of elements as the problem (represented by A above) has columns. If x is not present, a right hand side is contructed so that the solution is 1, 2, 3, ...m.

Input Parameter: x - the vector in the equation ``Ax = rhs"" mType - matrix type (optional). If the matrix is symmetric and only the lower or upper triangle is present, the user must let the routine know this by setting mType to one of: "L" or "l" - when only the lower triangle is present "U" or "u" - when only the upper triangle is present "T" or "t" - when either the lower or upper triangle is present. Updated Parameter: p - the problem for which the RHS is being constructed.

Sparspak.SpkProblem.outsparse — Method

outsparse(p::Problem{IT,FT})  where {IT,FT}

Output the sparse matrix.

Sparse LU Method

Sparspak.SpkSparseBase.factor — Method

factor(s::SparseSolver{IT}) where {IT}

Numerical factorization of the coefficient matrix.

Sparspak.SpkSparseBase.findorder — Method

findorder(s::SparseSolver{IT}, orderfunction::F) where {IT, F}

Find reordering of the coefficient matrix.

Sparspak.SpkSparseBase.findorder — Method

findorder(s::SparseSolver{IT}) where {IT, F}

Find reordering of the coefficient matrix using the default method.

Sparspak.SpkSparseBase.inmatrix — Method

inmatrix(s::SparseSolver{IT}, p::Problem{IT}) where {IT}

Put numerical values of the matrix stored in the problem into the data structures of the solver.

Sparspak.SpkSparseBase.symbolicfactor — Method

symbolicfactor(s::SparseSolver{IT})

Symbolic factorization of the(reordered) matrix A.

Create the data structures for the factorization and forward and backward substitution.

Sparspak.SpkSparseBase.triangularsolve — Method

triangularsolve(s::SparseSolver{IT, FT}, solution::Vector{FT}) where {IT, FT}

Forward and backward substitution (triangular solution).

Variant where the right-hand side vector is passed in.

Sparspak.SpkSparseBase.triangularsolve — Method

triangularsolve(s::SparseSolver{IT},  p::Problem{IT}) where {IT}

Forward and backward substitution (triangular solution).

Sparspak.SpkSparseSolver.findorderperm — Method

findorderperm(s::SparseSolver{IT}, perm) where {IT}

Find reordering of the coefficient matrix using a given permutation.

Sparspak.SpkSparseSolver.solve — Method

solve(s::SparseSolver{IT}, p::Problem{IT}) where {IT}

Execute all the steps of the solution process.

Given a symmetric matrix A, the steps are:

Reordering of the matrix A.

Symbolic factorization of the(reordered) matrix A.

Putting numerical values of A into the data structures.

Numerical factorization of A.

Forward and backward substitution (triangular solution).

The solution can be retrieved as p.x.

Graphs

Sparspak.SpkGraph.sortgraph — Method

SortGraph - sort the adjacency lists of the graph Important assumption: This works only for graphs that are symmetric. Output: updated graph

Ordering

Multiple minimum degree (MMD) ordering.

Sparspak.SpkMmd.generalmmd — Method

generalmmd(n, xadj, adj, perm, invp)

This routine implements the minimum degree algorithm. It makes use of the implicit representation of elimination graphs by quotient graphs, and the notion of indistinguishable nodes. It also implements the modifications by multiple elimination and minimum external degree.

Input parameters -

n - number of equations (xadj, adj) - adjacency structure for the graph. delta - tolerance value for multiple elimination. FIX ME: should delta be passed as argument?

Output: perm, invp - the minimum degree Ordering.

Working arrays - degHead (deg) - points to first node with degree deg, or 0 if there are no such nodes. degNext (node) - points to the next node in the degree list associated with node, or 0 if node was the last in the degree list. degPrev (node) - points to the previous node in a degree list associated with node, or the negative of the degree of node (if node was the last in the degree list), or 0 if the node is not in the degree lists. superSIze - the size of the supernodes. elimHead - points to the first node eliminated in the current pass Using elimNext, one can determine all nodes just eliminated. elimNext (node) - points to the next node in a eliminated supernode or 0 if there are no more after node. marker - marker vector. mergeParent - the parent map for the merged forest. needsUpdate (node) - > 0 iff node needs degree update.(0 otherwise)

Sparspak.SpkMmd.mmdelim — Method

Purpose - This routine eliminates the node mdNode of
  minimum degree from the adjacency structure, which
  is stored in the quotient Graph format.  It also
  transforms the quotient Graph representation of the
  elimination Graph.

Input parameters - mdNode - node of minimum degree. tag - tag value. invp - the inverse of an incomplete minimum degree Ordering. (It is zero at positions where the Ordering is unknown.) Updated parameters - (xadj, adjncy) - updated adjacency structure (xadj is not updated). degHead (deg) - points to first node with degree deg, or 0 if there are no such nodes. degNext (node) - points to the next node in the degree list associated with node, or 0 if node was the last in the degree list. degPrev (node) - points to the previous node in a degree list associated with node, or the negative of the degree of node (if node was the last in the degree list), or 0 if the node is not in the degree lists. superSIze - the size of the supernodes. elimNext (node) - points to the next node in a eliminated supernode or 0 if there are no more after node. marker - marker vector. mergeParent - the parent map for the merged forest. needsUpdate (node) - > 0 iff node needs update. (0 otherwise)

Sparspak.SpkMmd.mmdupdate — Method

Purpose - This routine updates the degrees of nodes
  after a multiple elimination step.

Input parameters - elimHead - the beginning of the list of eliminated nodes (i.e., newly formed elements). neqns - number of equations. (xadj, adjncy) - adjacency structure. delta - tolerance value for multiple elimination. invp - the inverse of an incomplete minimum degree Ordering. (It is zero at positions where the Ordering is unknown.) Updated parameters - mindeg - new minimum degree after degree update. degHead (deg) - points to first node with degree deg, or 0 if there are no such nodes. degNext (node) - points to the next node in the degree list associated with node, or 0 if node was the last in the degree list. degPrev (node) - points to the previous node in a degree list associated with node, or the negative of the degree of node (if node was the last in the degree list), or 0 if the node is not in the degree lists. superSIze - the size of the supernodes. elimNext (node) - points to the next node in a eliminated supernode or 0 if there are no more after node. marker - marker vector for degree update. tag - tag value. mergeParent - the parent map for the merged forest. needsUpdate (node) - > 0 iff node needs update. (0 otherwise)

Elimination Trees

Sparspak.SpkETree.binarytree — Method

To determine the binary tree representation of the elimination tree given by the parent vector. The returned representation will be given by the first - son and brother vectors. The root of the binary tree is always n.

Input Parameters: n - number of equations. parent - the parent vector of the elimination tree. It is assumed that parent(i) > i except for the roots.

Output Parameters: fson - the first son vector. brother - the brother vector.

Sparspak.SpkETree.findetree — Method

To determine the elimination tree from a given ordering and the adjacency structure. The parent vector is returned.

Input Parameters: n - number of equations. (xadj, adj) - the adjacency structure. (rPerm, rInvp) - permutation and inverse permutation vectors

Output Parameters: parent - the parent vector of the elimination tree.

Working Storage: ancestor - the ancestor vector.

Sparspak.SpkETree.getetree — Method

Given a graph and an ordering, GetETree finds the corresponding elimination tree. It calls the subroutine FindETree, which actually does the work.

Input Parameter: g - the graph whose elimination tree is to be found. order - the ordering for g

Updated Parameters: t - the elimination tree.

Sparspak.SpkETree.getpostorder — Method

GetPostorder finds a postordering of elimination tree t. The resulting ordering is returned in the object $order"". The vector argument$weight"" is optional. If it is present, the postordering will be one where the child vertices are ordered in increasing order of their weights. The elimination tree is reordered according to the postordering.

Input Parameters: t - the e - tree for which the postordering is found. weight - an optional weighting on the tree

Updated Parameters: order - the required ordering

Sparspak.SpkETree.postordertree — Method

Based on the binary representation (first - son, brother) of the elimination tree, a postordering is determined. The corresponding parent vector is also modified to reflect the reordering.

Input Parameters: root - root of the elimination tree (usually n). fson - the first son vector. brother - the brother vector.

Updated Parameters: parent - the parent vector.

Output Parameters: rInvpos - inverse permutation for the postordering.

Working Parameters: stack - the stack for postorder traversal of the tree.

Sparspak.SpkETree.weightedbinarytree — Method

To determine a binary tree representation of the elimination tree, for which every "last child" has the maximum possible column nonzero count in the factor. The returned representation will be given by the first - son and brother vectors. The root of the binary tree is always n.

Input Parameters: n - number of equations. parent - the parent vector of the elimination tree. It is assumed that parent(i) > i except for the roots. weight - a weighting on the tree.

Output Parameters: fson - the first son vector. brother - the brother vector.

Working Storage: lson - last son vector.

Symbolic Factorization

Grid